SRE-3023 Cleaning up is more drastic

+++ spinner/clean.sh
@@ -1,5 +1,9 @@
 #!/bin/bash
 
+function lcd {
+	cd "${1}" || exit 3
+}
+
 docker_compose="docker compose"
 
 if (command -v docker-compose &>/dev/null)
@@ -7,14 +11,21 @@ then
 	docker_compose="docker-compose"
 fi
 
-for environment in $(ls | grep "env-")
+base_dir=$(pwd)
+
+if [[ $(find . -name "env-*" -type d | wc -l) -eq 0 ]]
+then
+	echo "There were no environments to clean."
+
+	exit 2
+fi
+
+for environment in "$(pwd)"/env*
 do
 	echo "Cleaning up ${environment}."
 
-	cd "${environment}"
-	${docker_compose} down
-
-	cd ..
+	lcd "${environment}"
+	${docker_compose} down --local all -v
 
 	rm -fr "${environment}"
 done
